lots of speedups
1. generalised FastVector, use it for stuff
2. use int32_t for Parser::Pin to reduce memory footprint (64KB will be enough, surely! - 1980s)
3. reduced copies everywhere as much as possible by maintaining immutability
4. better import AST finding that has the lexer return a list of indices in the token list (which, did i mention, is immutable)
5. tweak the weird process library (unix at least) to not make `sh` launch `sh` again…
6. support multi-line strings through the trailing backslash thing

+++ source/include/defs.h
@@ -47,12 +47,12 @@ namespace Parser
 	struct Pin
 	{
 		Pin() : fileID(0), line(1), col(1), len(0) { }
-		Pin(size_t f, size_t l, size_t c, size_t ln) : fileID(f), line(l), col(c), len(ln) { }
+		Pin(int32_t f, int32_t l, int32_t c, int32_t ln) : fileID(f), line(l), col(c), len(ln) { }
 
-		size_t fileID;
-		size_t line;
-		size_t col;
-		size_t len;
+		int32_t fileID;
+		int32_t line;
+		int32_t col;
+		int32_t len;
 	};
 
 	struct Token;
